Implemented new structures to allow for the Player to carry a melee cone.
The plan is to use this cone to help detect entities in front of player and determine who the player damages when they use their melee weapon. Currently, it draws the melee cone of the "fists" weapon and draws it on the screen. It successfully updates it's position and orientation corresponding to the status of the player and mouse.

+++ Headers/weaponStats.h
@@ -31,7 +31,7 @@ private:
 
 	int meleeDamage[TOTAL_MELEE_TYPES];
 	double meleeConeOfEffectAngle[TOTAL_MELEE_TYPES];
-	int meleeRange[TOTAL_MELEE_TYPES];
+	float meleeRange[TOTAL_MELEE_TYPES];
 	int meleePrice[TOTAL_MELEE_TYPES];
 	int meleeKnockback[TOTAL_MELEE_TYPES];
 	int meleeSwingRate[TOTAL_MELEE_TYPES];
@@ -54,7 +54,7 @@ public:
 	// Get the angle of the cone of effect for the specified melee type in degrees.
 	double getMeleeAngle(int weapon);
 	// Get the range of the cone of effect for the specified melee type in pixels.
-	int getMeleeRange(int weapon);
+	float getMeleeRange(int weapon);
 	// Get the price of the specified melee type.
 	int getMeleePrice(int weapon);
 	// Get the knockback power of the specified melee type in the number of pixels the target would be knocked back by.
